Small refactors
made some small refactors , deleted the JWTController and added AuthenticationController with proper urls.
